MLFLOW/api/requirements.txt
fastapi
uvicorn[standard]
mlflow-skinny>=2.0.0,<3.0.0 # Pin mlflow version range if desired
scikit-learn
xgboost
pandas
numpy